LIME-138: * Edit models, seeds and rename activityType
