+++ backend/src/bundles/achievements/achievement.model.ts
@@ -1,18 +1,34 @@
+import { type RelationMappings, Model } from 'objection';
+
 import {
     AbstractModel,
     DatabaseTableName,
 } from '~/common/database/database.js';
 
 import { type ActivityType, type Metric, type ValueOf } from './enums/enums.js';
+import { UserAchievementModel } from './user-achievement.model.js';
 
 class AchievementModel extends AbstractModel {
     public 'name': string;
-    public 'activity': ValueOf<typeof ActivityType> | null;
+    public 'activityType': ValueOf<typeof ActivityType> | null;
     public 'requirement': number;
     public 'requirementMetric': ValueOf<typeof Metric> | null;
 
     public static override get tableName(): string {
         return DatabaseTableName.ACHIEVEMENTS;
+    }
+
+    public static override get relationMappings(): RelationMappings {
+        return {
+            userAchievement: {
+                relation: Model.HasOneRelation,
+                modelClass: UserAchievementModel,
+                join: {
+                    from: `${DatabaseTableName.ACHIEVEMENTS}.id`,
+                    to: `${DatabaseTableName.USER_ACHIEVEMENTS}.achievementId`,
+                },
+            },
+        };
     }
 }
