+++ backend/src/seeds/enums/enums.ts
@@ -1,0 +1,1 @@
+export { ActivityType, Metric } from 'shared';
